feat: add popup for log out
