+++ src/components/ModalContentInProfile/ModalContentInProfile.scss
