+++ src/pages/ProfilePage/ProfilePage.tsx
@@ -1,6 +1,6 @@
 import { ReactNode, useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { useAppSelector } from "../../app/hooks";
+import { useAppDispatch, useAppSelector } from "../../app/hooks";
 import Footer from "../../components/Footer/Footer";
 import Header from "../../components/Header/Header";
 import "./ProfilePage.scss";
@@ -8,16 +8,22 @@ import ProfileGeneral from "../../components/ProfileGeneral/ProfileGeneral";
 import ProfileChangePassword from "../../components/ProfileChangePassword/ProfileChangePassword";
 import Modal from "../../components/Modal/Modal";
 import useModal from "../../hooks/useModal";
-import DeleteAccount from "../../components/DeleteAccount/DeleteAccount";
+import ModalContentInProfile from "../../components/ModalContentInProfile/ModalContentInProfile";
 import logOut from "../../API/logOut";
+import DeleteAccountImg from "../../assets/png/delete_account.png";
+import LogOutImg from "../../assets/png/log_out.png";
+import deleteAccount from "../../API/deleteAccount";
+import { SnackbarTextValue } from "../../features/SnackbarTextSlice";
 
 type ProfileStates = "general" | "changePassword";
 
 export default function ProfilePage() {
   const isAuth = useAppSelector((state) => state.present.authData.value.isAuth);
   const name = useAppSelector((state) => state.present.authData.value.name);
+  const token = useAppSelector((state) => state.present.authData.value.token);
   const isOpen = useAppSelector((state) => state.present.IsOpenModal.value);
   const navigate = useNavigate();
+  const dispatch = useAppDispatch();
   const [profileState, SetProfileState] = useState<ProfileStates>("general");
   const { toggle } = useModal();
 
@@ -68,8 +74,17 @@ export default function ProfilePage() {
                         type="button"
                         className="profile-menu__btn"
                         onClick={() => {
-                          navigate("/");
-                          logOut(name);
+                          toggle();
+                          setModalChildren(
+                            <ModalContentInProfile
+                              imageSrc={LogOutImg}
+                              text="Are you sure you want to log out?"
+                              handleClickByOkBtn={() => {
+                                navigate("/");
+                                logOut(name);
+                              }}
+                            />
+                          );
                         }}
                       >
                         Log Out
@@ -80,7 +95,24 @@ export default function ProfilePage() {
                       className="profile-menu__btn"
                       onClick={() => {
                         toggle();
-                        setModalChildren(<DeleteAccount />);
+                        setModalChildren(
+                          <ModalContentInProfile
+                            imageSrc={DeleteAccountImg}
+                            text="Are you sure you want to delete your account?"
+                            handleClickByOkBtn={() => {
+                              deleteAccount(token).then(() => {
+                                toggle();
+                                navigate("/");
+                                dispatch(
+                                  SnackbarTextValue({
+                                    text: "Your account was deleted",
+                                    withUndo: false,
+                                  })
+                                );
+                              });
+                            }}
+                          />
+                        );
                       }}
                     >
                       Delete account
